@@ -11,6 +11,7 @@
     "start": "yarn storybook",
     "test": "BABEL_ENV=test jest",
     "test:watch": "BABEL_ENV=test jest --watch",
+    "lint": "eslint ./src ./__tests__",
     "prettier": "prettier --tab-width 4 --single-quote --trailing-comma es5 --print-width 80 --write \"{src,examples,__tests__}/**/**/**/*.js\"",
     "prepublish": "babel ./src --out-dir ./lib --presets es2015,react,stage-0 --plugins babel-plugin-add-module-exports",
     "storybook": "start-storybook -p 3000",
@@ -21,6 +22,9 @@
   "jest": {
     "modulePaths": [
       "<rootDir>/src/"
+    ],
+    "testMatch": [
+      "**/?(*.)+(spec|test).[jt]s?(x)"
     ],
     "coverageDirectory": "./coverage/",
     "collectCoverage": true
@@ -49,6 +53,7 @@
     "@storybook/react": "^3.3.9",
     "babel-cli": "^6.24.1",
     "babel-core": "^6.23.1",
+    "babel-eslint": "^10.0.1",
     "babel-jest": "^20.0.3",
     "babel-loader": "^6.3.2",
     "babel-plugin-add-module-exports": "^0.2.1",
@@ -59,13 +64,17 @@
     "codecov": "^3.0.0",
     "cross-env": "^3.1.4",
     "css-loader": "^0.28.0",
+    "eslint": "^5.13.0",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-jest": "^22.2.2",
+    "eslint-plugin-react": "^7.12.4",
     "express": "^4.14.1",
     "jest": "22.0.6",
     "node-sass": "^4.5.0",
     "postcss-loader": "^1.3.3",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-test-renderer": "^16.4.1",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
+    "react-test-renderer": "^16.8.0",
     "regenerator-runtime": "^0.10.5",
     "sass-loader": "^6.0.3",
     "style-loader": "^0.16.1",
