@@ -1,3 +1,2 @@
 // Addons go here
-
-// import '@storybook/addon-knobs/register';
+import '@storybook/addon-knobs/register';
