@@ -2,18 +2,15 @@
 
 import React from 'react';
 
-import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
-import ReactDOMServer from 'react-dom/server';
 import ParallaxProvider from 'components/ParallaxProvider';
 import ParallaxController from 'modules/ParallaxController';
 import withController from 'components/withController';
 
 describe('A <ParallaxProvider>', () => {
-    // afterEach(() => )
     it('to render children', () => {
         const node = document.createElement('div');
-        let child = jest.fn();
+        const child = jest.fn();
         const Child = () => {
             child();
             return <div />;
@@ -72,29 +69,6 @@ describe('A <ParallaxProvider>', () => {
         expect(spy).toBeCalled();
     });
 
-    it('to copy the instance to the window for legacy versions', () => {
-        const node = document.createElement('div');
-        ReactDOM.render(
-            <ParallaxProvider>
-                <div />
-            </ParallaxProvider>,
-            node
-        );
-        expect(window.ParallaxController).toBeInstanceOf(ParallaxController);
-    });
-
-    it('to remove the global instance on the window for legacy versions', () => {
-        const node = document.createElement('div');
-        ReactDOM.render(
-            <ParallaxProvider>
-                <div />
-            </ParallaxProvider>,
-            node
-        );
-        ReactDOM.unmountComponentAtNode(node);
-        expect(window.ParallaxController).toBeNull();
-    });
-
     it('to always create a new instance when re-mounting', () => {
         // the provider isn't gauranteed to be destroyed before re-instantiated
         // in a route change.
@@ -117,18 +91,14 @@ describe('A <ParallaxProvider>', () => {
 
         // first instance mounted
         const instance1 = render(node1);
-        expect(window.ParallaxController).toBeInstanceOf(ParallaxController);
         expect(instance1.controller).toBeInstanceOf(ParallaxController);
 
         // second instance mounted
         const instance2 = render(node2);
-        expect(window.ParallaxController).toBeInstanceOf(ParallaxController);
         expect(instance2.controller).toBeInstanceOf(ParallaxController);
 
         // unmount first instance
         ReactDOM.unmountComponentAtNode(node1);
-        // this is ok because the global is only required by <= 1.1.0, and a provider isn't used.
-        expect(window.ParallaxController).toBeNull();
 
         // this must still be defined
         expect(instance2.controller).toBeInstanceOf(ParallaxController);
